move ES functions to common lib
